python/requirements.txt
Pillow>=10.0.0
matplotlib>=3.9.0
numpy>=2.0.0

